Finish to implement the start script in nexus-scripts package
