+++ packages/nexus-scripts/lib/scripts/eject/index.js
@@ -1,0 +1,3 @@
+const eject = require('./eject');
+
+module.exports = eject;
