+++ packages/nexus-scripts/lib/scripts/utils/clearConsole.js
@@ -1,0 +1,7 @@
+function clearConsole() {
+  process.stdout.write(
+    process.platform === 'win32' ? '\x1B[2J\x1B[0f' : '\x1B[2J\x1B[3J\x1B[H',
+  );
+}
+
+module.exports = clearConsole;
