+++ packages/nexus-scripts/lib/scripts/webpack/configurations/webpack.prod.js
