+++ packages/nexus-scripts/package.json
@@ -38,6 +38,7 @@
     "@babel/register": "7.4.4",
     "@babel/runtime": "7.4.4",
     "babel-loader": "8.0.5",
+    "case-sensitive-paths-webpack-plugin": "2.2.0",
     "chalk": "2.4.2",
     "circular-dependency-plugin": "5.0.2",
     "core-js": "3",
@@ -45,6 +46,7 @@
     "express": "4.16.4",
     "html-webpack-plugin": "3.2.0",
     "ip": "1.1.5",
+    "open": "6.2.0",
     "regenerator-runtime": "0.13.2",
     "terser-webpack-plugin": "1.2.3",
     "webpack": "4.30.0",
